Added a common generic interface for all direct linear solvers (DLS) aling the same lines as the SPILS interface for the iterative solvers. - both the internal direct linear solvers (band and dense), as well   as the Blas/Lapack ones (both band and dense) build on top of this   common interface. - there is only one type for matrices, namely DlsMat, which has   a few fields that are relevant only for banded matrices. - all dimensions (problem size, half-bandwidths, etc) are now   of type int (instead of long int) - the "smalldense" generic linear solver has been combined into   the "dense" one, similar to what we already had for "band" - the allocation and deallocation functions now defined in   sundials_direct.h were renamed for a more "consistent" scheme   across SUNDIALS (e.g. NewDenseMat, DestroyMat) - added a couple of examples using the Lapack linear solvers   (only for CVODE and CVODES at this time; more later)
- updated all FCMIX interfaces
- updated all examples
- updated sundialsTB

- documentation NOT updated yet

+++ examples/cvode/parallel/cvkryx_bbd_p.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.1 $
- * $Date: 2006-07-05 15:50:05 $
+ * $Revision: 1.2 $
+ * $Date: 2006-11-22 00:12:44 $
  * -----------------------------------------------------------------
  * Programmer(s): S. D. Cohen, A. C. Hindmarsh, M. R. Wittman, and
  *                Radu Serban  @ LLNL
@@ -151,8 +151,8 @@ static int f(realtype t, N_Vector u, N_Vector udot, void *f_data);
 
 /* Prototype of functions called by the CVBBDPRE module */
 
-static int flocal(long int Nlocal, realtype t, N_Vector u,
-                   N_Vector udot, void *f_data);
+static int flocal(int Nlocal, realtype t, N_Vector u,
+                  N_Vector udot, void *f_data);
 
 /* Private function to check function return values */
 
@@ -168,7 +168,7 @@ int main(int argc, char *argv[])
   realtype abstol, reltol, t, tout;
   N_Vector u;
   int iout, my_pe, npes, flag, jpre;
-  long int neq, local_N, mudq, mldq, mukeep, mlkeep;
+  int neq, local_N, mudq, mldq, mukeep, mlkeep;
   MPI_Comm comm;
 
   data = NULL;
@@ -722,7 +722,7 @@ static int f(realtype t, N_Vector u, N_Vector udot, void *f_data)
    inter-processor communication of data needed to calculate f has already
    been done, and this data is in the work array uext.                    */
 
-static int flocal(long int Nlocal, realtype t, N_Vector u,
+static int flocal(int Nlocal, realtype t, N_Vector u,
                   N_Vector udot, void *f_data)
 {
   realtype *uext;
